Update device-restrictions-ios.md
we have done extensive testing and we can observe the behavior. if additional information is needed please contact me on ulrugwir and i will provide ICM links or testing scenario.

added: on Line 431

Note: "3rd party VPN applications might be able to add the VPN profiles even when "Allow adding VPN Configuration (Supervised Only) /allowVPNCreation" is configured on the device. This is an iOS behavior"